+++ app-main/src/test/java/net/causw/app/main/domain/integration/crawled/service/CrawlServiceTest.java
@@ -2,9 +2,11 @@ package net.causw.app.main.domain.integration.crawled.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.verify;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -13,6 +15,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
+import net.causw.app.main.domain.community.post.entity.Post;
+import net.causw.app.main.domain.community.post.service.implementation.PostWriter;
 import net.causw.app.main.domain.integration.crawled.SiteConfigFixture;
 import net.causw.app.main.domain.integration.crawled.core.CrawlContext;
 import net.causw.app.main.domain.integration.crawled.core.SiteCrawlerRegistry;
@@ -22,7 +26,10 @@ import net.causw.app.main.domain.integration.crawled.dto.CleanArticle;
 import net.causw.app.main.domain.integration.crawled.dto.CrawlResult;
 import net.causw.app.main.domain.integration.crawled.dto.CrawlSaveStatus;
 import net.causw.app.main.domain.integration.crawled.dto.RawArticle;
+import net.causw.app.main.domain.integration.crawled.entity.CrawledNotice;
 import net.causw.app.main.domain.integration.crawled.entity.SiteConfig;
+import net.causw.app.main.domain.integration.crawled.service.implementation.CrawledNoticeReader;
+import net.causw.app.main.domain.integration.crawled.service.implementation.CrawledNoticeWriter;
 import net.causw.app.main.domain.integration.crawled.service.implementation.SiteConfigReader;
 
 @ExtendWith(MockitoExtension.class)
@@ -38,7 +45,11 @@ class CrawlServiceTest {
 	@Mock
 	private CrawledArticleCleaner cleaner;
 	@Mock
-	private CrawledNoticeUpsertService crawledNoticeUpsertService;
+	private CrawledNoticeReader crawledNoticeReader;
+	@Mock
+	private CrawledNoticeWriter crawledNoticeWriter;
+	@Mock
+	private PostWriter postWriter;
 	@Mock
 	private SiteCrawler crawler;
 
@@ -61,7 +72,7 @@ class CrawlServiceTest {
 		given(crawler.fetchArticle(context, failed)).willThrow(new IllegalStateException("failed"));
 		given(crawler.fetchArticle(context, succeeded)).willReturn(raw);
 		given(cleaner.clean(raw, config)).willReturn(clean);
-		given(crawledNoticeUpsertService.upsert(clean)).willReturn(CrawlSaveStatus.CREATED);
+		given(crawledNoticeWriter.upsert(clean)).willReturn(CrawlSaveStatus.CREATED);
 
 		// when
 		CrawlResult result = crawlService.crawl("site");
@@ -69,6 +80,42 @@ class CrawlServiceTest {
 		// then
 		assertThat(result.createdCount()).isEqualTo(1);
 		assertThat(result.failedUrls()).containsExactly(failed.url());
+	}
+
+	@Test
+	@DisplayName("대상 게시판이 변경되면 연결된 Post를 삭제하고 공지를 갱신한다")
+	void crawl_shouldSoftDeleteLinkedPost_whenTargetBoardChanges() {
+		// given
+		ArticleUrl articleUrl = new ArticleUrl("https://example.com/10", "10", "공지");
+		SiteConfig config = config();
+		CrawlContext context = new CrawlContext(config);
+		RawArticle raw = new RawArticle("site", "10", articleUrl.url(), "공지", "제목", "본문", "관리자",
+			"2026-08-10", null, List.of());
+		CleanArticle article = new CleanArticle("site", "target-board-id", "10", articleUrl.url(), "공지", "제목", "본문",
+			"관리자",
+			LocalDate.of(2026, 8, 10), null, List.of(), "hash");
+		CrawledNotice notice = CrawledNotice.of(
+			"site", "10", "old-board-id", "공지", "제목", "본문", articleUrl.url(), "관리자",
+			LocalDate.of(2026, 8, 10), null, List.of(), "old-hash");
+		Post post = org.mockito.Mockito.mock(Post.class);
+		notice.linkPost(post);
+
+		given(siteConfigReader.getEnabledBySiteId("site")).willReturn(config);
+		given(registry.get(config.getCrawlerType())).willReturn(crawler);
+		given(crawler.fetchList(context)).willReturn(List.of(articleUrl));
+		given(crawler.fetchArticle(context, articleUrl)).willReturn(raw);
+		given(cleaner.clean(raw, config)).willReturn(article);
+		given(crawledNoticeReader.findBySource("site", "10")).willReturn(Optional.of(notice));
+		given(crawledNoticeWriter.upsert(article)).willReturn(CrawlSaveStatus.UPDATED);
+
+		// when
+		CrawlResult result = crawlService.crawl("site");
+
+		// then
+		assertThat(result.updatedCount()).isEqualTo(1);
+		verify(post).setIsDeleted(true);
+		verify(postWriter).save(post);
+		verify(crawledNoticeWriter).upsert(article);
 	}
 
 	@Test
